Add new single demo script for all ISO 15118-2 AC with OCPP demos
This new script includes option flags for which demo to run.

It also enables the option to test a user's local copy of this repo
without having to edit the script or push commits up to GitHub first.

Note that the OCPP v1.6j demo is currently not working with this new
script (though it also wasn't working with the old script either).

Signed-off-by: Bryan T. Richardson <bryan@activeshadow.com>