Changed branches to run workflow on
Not running on main since I will be making code pushes to my feature-branch: automate-tests-actions.
Also, created a new branch automate-tests-merge that will be used to test workflow runs whenever a PR is created.
This will simulate creating a PR for merging code changes into main in the parent repo.

Signed-off-by: Mahadik, Mukul Chandrakant <MukulChandrakant.Mahadik@nrel.gov>