Add a step to print outputs from docker setup GitHub action for MacOS
Previous run failed yet again at the docker-verify state with error:

```
  shell: /bin/bash --noprofile --norc -e -o pipefail {0}
/Users/runner/work/_temp/f2669228-43bc-4f11-9c38-82bd32c1e3b4.sh: line 1: docker: command not found
Error: Process completed with exit code 127.
```

Hence printing outputs as mentioned in the action Readme.md.

-----

Ah! I see the step to install from the docker action was skipped as I had missed changing the macos version to v 13 in the if condition.
Let's see if it works now.

Signed-off-by: Mahadik, Mukul Chandrakant <MukulChandrakant.Mahadik@nrel.gov>