Using artifacts to make mqtt server image accessible in manager matrix job
Found out an issue with the latest tagged image not being fetched in the docker compose command.
The reason was that the matrix job only loaded its respective image from the cache.
But the docker compose command for running automated tests uses both manager and mqtt server images.

To solve this, initially I used docker build push action to reload the mqtt image in the matrix job for manager.
But then finally went with the approach of using artifacts to share the mqtt image between jobs - uploaded in mqtt matrix job, downloaded in manager matrix job.

Signed-off-by: Mahadik, Mukul Chandrakant <MukulChandrakant.Mahadik@nrel.gov>