♻️ 🔒️Validate in the lower level + add validation to Citrine
- Move the validation to the lower level since the directory
  structure is different
- Add code to copy the real root over to citrine
- Add validation to citrine

As an aside, I wonder if this weird configuration (the actual
root CA isn't there?!) is the reason that the self-signed
certs are not working with Citrine. Need to investigate once
we roll forward.

Testing done:
as part of 86b1ea8ae7cbd1e8beaf7a42f53f8d752ec2e967

Signed-off-by: Shankari <k.shankari@ee.doe.gov>