🗃️ Change the way that the CSMS URL is specified
Now that the device model DB is overridden on startup,
we cannot configure the CSMS URL by copying over the database
any more (b613ba6c49f00339a090abe033aecfb45101dbad,
https://github.com/EVerest/everest-demo/issues/78#issuecomment-2477536249).

Instead we have to edit the `InternalCtrlr.json` file before the server starts up so that it is configured properly in the device model DB that is created at startup. We do this by using `sed` to replace `localhost` with the correct CSMS URL.

But the CSMS URL is different based on the the profile (SP1,
SP2, SP3) and the CSMS (Maeve vs CitrineOS).

So we make the following changes:
- remove all the DB override files since they are not
  relevant any more
- instead, define the SP and CSMS specific enviroment
  variables in the CSMS-specific apply-patches file
    - we may want to change this to a separate file later if
      there are other environment variables we want to
      specify
    - specifying this in CSMS-specific files means that we
      can easily support other CSMSes without modifying the
      demo file
- change the demo script to run separate sed commands that
  modify the `InternalCtrlr.json` with the correct URL for the service profile and CSMS

Signed-off-by: Shankari <k.shankari@ee.doe.gov>