Don't override the AC flow
Now that we have made all the changes to the checked-in flow,
we don't need to copy it over any more. Hopefully, this will
also encourage us to commit code to main more often and not
work for too long off branches on forks.

Testing done: None, but it is a pretty obvious change

Signed-off-by: Shankari <k.shankari@driveelectric.gov>