🔧 Configure limits in a way that makes sense to me
It looks like the limits applied to the JsYetiSimulator are
on a per-phase basis, although limits applied elsewhere are
not. This makes things very confusing. We will fix this by
configuring everything to a per-phase limit of 16A

Please see
https://github.com/EVerest/everest-demo/issues/92#issuecomment-2484715367
to
https://github.com/EVerest/everest-demo/issues/92#issuecomment-2485076808

Locations changed:
- config
- JsYetiSimulator
- SmartCharging OCPP defaults

Also fix the disable_iso_tls patch to not have a starting `/`

Add a new patch file to enable limit logging but don't enable
it by default. This may help future efforts to debug this.

Signed-off-by: Shankari <k.shankari@driveelectric.gov>