💅 Change the text on the CSMS override UI to be more meaningful
To be consistent with OCPP 2.0.1, the "External Limits" should be used for a
local EMS or other controller. However, we have the CSMS now send limits from
two sources - MIDAS or the CSMS' internal operation. Both of them are sent as
`ChargingStationMaxProfile`. Changing the "CSMS internal operation" label to be
more meaningful.

Signed-off-by: Shankari <k.shankari@driveelectric.gov>