💻 "dev" environment: edit code locally; run in a container
As more folks from US-JOET start working on everest, and we start making
changes that are not just "hacks", we are running into issues because of the
current process for making changes. This process involves launching the
containers and then editing the code in them directly using vim. This is slow
and painful, requires manually copy-pasting commands for the final
configuration, and does not allow developers to use modern IDEs.

This refactors the docker process to pull out the setup and compile functions
into separate scripts. The scripts are then used by the Dockerfile for the
pre-build images, but can also be run by a developer to recompile and rebuild.

This also creates a "dev" docker-compose, which mounts the specified source
directory as a volume, allowing us to edit it on the host using more
sophisticated editors.

Finally, it edits the README to provide the steps for a developer to use this new functionality.

Testing done:
- Re-built manager images, re-launched, was able to start up manager
- Deployed dev image, ran steps in README, compiled, was able to start up manager

Signed-off-by: Shankari <k.shankari@driveelectric.gov>