🔨 Change all instances of `patch` to `patch -N`
This hasn't been an issue with the pre-built container, but with the dev
containers, we will likely apply the same patches over and over again using the
setup scripts, on the same codebase. The second time, the patches would already
have been applied.

Passing in `patch -N` (aka --forward) skips reverse patches instead of pausing
and asking for a resolution.

Testing done:
- Ran the script in the dev environment without errors; started the manager after install

Signed-off-by: Shankari <k.shankari@driveelectric.gov>