Support specifying the charge station as an input parameter
Before this, the charge station was hardcoded as `cp001`, which was convenient
for testing, but made running in more complex scenarios. In particular, with
the recent addition of the "dev" environment, it may be useful to run dev and
prebuilt environments side-by-side, so that we can verify behavior with and
without changes.

This commit plumbs through the charge station in both dev and prebuilt
environments.
- In the dev environment, the `CHARGE_STATION_ID` is in the `docker-compose`
- In the pre-built environment, it is passed in as a command line argument,
  defaulting to `cp001`

Note that in the `dev` environment, the new `CHARGE_STATION_ID` needs to be
added manually using the `DEMO_VERSION=sp1 CHARGE_STATION_ID=cpdev bash maeve/add-charger-and-rfid-card.sh` command.

The README has been updated to clarify these instructions.

If run through the `demo-XXX.sh` script, the script will automatically add the station.

Testing done:
- Ran the dev environment with station `cpdev`; manager was able to connect to the CSMS
- Ran the `demo` script with station `cpdemo`; manager was able to connect to the CSMS

Signed-off-by: Shankari <k.shankari@driveelectric.gov>