🚑️ Fix regression; fill in the `EVSENominalVoltage`
While implementing the departure time hack, we removed the nominal voltage
population because we already had the nom_voltage but did not put it in
anywhere.

This caused it to be populated with a default value of zero hours
`EVSENominalVoltage":{"Multiplier":0,"Unit":"h","Value":0}`

which failed validation.

Simply restoring that change and regenerating the diff fixes the issue

This fixes https://github.com/EVerest/everest-demo/issues/103

Signed-off-by: Shankari <k.shankari@driveelectric.gov>